fix: Use the persistence path value for Fluentd
This commit uses the Artifactory persistence mount path value for the
Artifactory log folder volume which is mounted in the Fluentd container.

It also removes a section from the Fluentd config which does not seem to
be needed and also doesn't work given it was written expecting that
Fluentd was installed directly into the same container/server where
Artifactory is running.