Avoid issue with the poststartcommand
The postStartCommand prevent me to run the chart version 2.5.7 because i receive this error message :
```shell
art-artifactory-ha-primary-0   0/1     PostStartHookError: command '/bin/bash -c echo;
' exited with 126:
```

The current version of the code alway execute */bin/bash -c echo;* during the post start hook even if we don't defined a command.  This PR which i did not test should prevent the hook to be executed if the user did not specify a command.